formally switch from roadblock.py to roadblocker.py
- ie. don't use symbolic links to fake the switch
